Added detail view with “Play Episode” button that will take user to the player view.
Changed the summary view to use a web view and display html content from the feed.  Links in the summary view will open up in Safari.

Fixed the centering of the Bad Voltage logo on the list view.

Resized the scrubber to avoid accidental “back” swipes (and to make room for labels indicating time left and time played).